Add server-side configurations and ring enchantment updates
This update introduces server-side configurations for ring enchantments. It also refactors how the "Susurrus" ring enchantment is applied. The ticks are now managed through a new RingComponent instead of via tags on the enchantment item stack. This simplifies the code and improves performance. The targets of enchantments are now triggered based on a configurable server-side setting. This provides a greater level of customization for server admins. Gradle variables are also now used in the pack.mcmeta and fabric.mod.json